+++ src/rhi/storage.hpp
@@ -21,7 +21,7 @@ class CommonStorage : public IStorage
     std::vector<ReadOnlyFilePtr> openFiles(const fs::path& path, const fs::path& ext) override;
 
     void requestLoadTexture(coro::latch& latch, BindlessTablePtr& table, const std::span<ReadOnlyFilePtr>& files) override;
-    void requestLoadBuffer(coro::latch& latch, const ReadOnlyFilePtr& file, BufferPtr& buffer, uint32_t fileLength, uint32_t fileOffset) override;
+    void requestLoadBuffer(coro::latch& latch, const ReadOnlyFilePtr& file, BufferPtr& buffer, uint64_t fileLength, uint64_t fileOffset) override;
 
     img::ITexture* factoryTexture(const ReadOnlyFilePtr& file, std::byte* metadata);
     const BufferPtr& getStaging(int index) const { return m_stagings[index]; }
@@ -39,7 +39,7 @@ class CommonStorage : public IStorage
   private:
     virtual coro::task<> makeSingleTextureTask(coro::latch& latch, BindlessTablePtr table, ReadOnlyFilePtr file) = 0;
     virtual coro::task<> makeMultiTextureTask(coro::latch& latch, BindlessTablePtr table, std::vector<ReadOnlyFilePtr> files) = 0;
-    virtual coro::task<> makeBufferTask(coro::latch& latch, ReadOnlyFilePtr file, BufferPtr buffer, uint32_t fileLength, uint32_t fileOffset) = 0;
+    virtual coro::task<> makeBufferTask(coro::latch& latch, ReadOnlyFilePtr file, BufferPtr buffer, uint64_t fileLength, uint64_t fileOffset) = 0;
 
     using task_container = coro::task_container<coro::thread_pool>;
     task_container m_scheduler;
