+++ src/rhi/vulkan/vulkan_storage.cpp
@@ -320,8 +320,8 @@ coro::task<> makeMultiTextureTaskDeprecated(coro::latch& latch, TexturePoolPtr t
     co_return;
 }*/
 
-coro::task<> Storage::makeBufferTask(coro::latch& latch, ReadOnlyFilePtr file, BufferPtr buffer, uint32_t fileLength,
-                                     uint32_t fileOffset)
+coro::task<> Storage::makeBufferTask(coro::latch& latch, ReadOnlyFilePtr file, BufferPtr buffer, uint64_t fileLength,
+                                     uint64_t fileOffset)
 {
     const int bufferId = acquireStaging();
     BufferPtr staging = getStaging(bufferId);
